+++ plugins/circleci/src/pages/BuildsPage/index.ts
@@ -1,0 +1,1 @@
+export * from './BuildsPage';
